feature: divide firebase.server.ts. functions to separate files
